@@ -1,8 +1,7 @@
 {
-  "name": "chapter9",
+  "name": "chapter12",
   "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
+  "description": "Chapter 12 of the WebdriverIO workshop: TypeScript",
   "scripts": {
     "test": "wdio wdio.local.conf.ts"
   },
@@ -10,16 +9,14 @@
   "author": "Christian Bromann <mail@bromann.dev>",
   "license": "MIT",
   "devDependencies": {
-    "@wdio/allure-reporter": "^7.20.3",
-    "@wdio/cli": "^7.20.4",
-    "@wdio/local-runner": "^7.20.4",
-    "@wdio/mocha-framework": "^7.20.3",
-    "@wdio/spec-reporter": "^7.20.3",
-    "allure-commandline": "^2.17.2",
-    "chromedriver": "^103.0.0",
-    "ts-node": "^10.8.1",
-    "typescript": "^4.7.4",
-    "wdio-chromedriver-service": "^7.3.2",
-    "webdriverio": "^7.20.4"
+    "@wdio/allure-reporter": "^8.34.0",
+    "@wdio/cli": "^8.35.1",
+    "@wdio/local-runner": "^8.35.1",
+    "@wdio/mocha-framework": "^8.35.0",
+    "@wdio/spec-reporter": "^8.32.4",
+    "allure-commandline": "^2.27.0",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.4.3",
+    "webdriverio": "^8.35.1"
   }
 }
