@@ -2,26 +2,13 @@ import allure from 'allure-commandline'
 
 import { config as sharedConfig } from './wdio.conf'
 
-export const config: WebdriverIO.Config = Object.assign(sharedConfig, {
-  capabilities: [{
-    // maxInstances can get overwritten per capability. So if you have an in-house Selenium
-    // grid with only 5 firefox instances available you can make sure that not more than
-    // 5 instances get started at a time.
-    maxInstances: 5,
-    //
-    browserName: 'chrome'
-    // If outputDir is provided WebdriverIO can capture driver session logs
-    // it is possible to configure which logTypes to include/exclude.
-    // excludeDriverLogs: ['*'], // pass '*' to exclude all driver session logs
-    // excludeDriverLogs: ['bugreport', 'server'],
-  }],
+export const config: WebdriverIO.Config = {
+  ...sharedConfig,
 
-  services: [
-    [
-      'chromedriver',
-      { outputDir: './logs' }
-    ]
-  ],
+  capabilities: [{
+    maxInstances: 5,
+    browserName: 'chrome'
+  }],
 
   /**
    * Gets executed after all workers got shut down and the process is about to exit.
@@ -30,7 +17,7 @@ export const config: WebdriverIO.Config = Object.assign(sharedConfig, {
    * @param {Array.<Object>} capabilities list of capabilities details
    * @param {<Object>} results object containing test results
    */
-  onComplete: function() {
+  onComplete: () => {
     const generation = allure(['generate', 'allure-results'])
     return new Promise<void>((resolve, reject) => {
       const generationTimeout = setTimeout(
@@ -44,4 +31,4 @@ export const config: WebdriverIO.Config = Object.assign(sharedConfig, {
       })
     })
   }
-})
+}
